added lists 2 and 3 of rust

+++ 2019-2020/Zima/Rust/List 1/ex4/src/main.rs
@@ -1,5 +1,5 @@
 fn square_area_to_circle(size:f64) -> f64 {
-    std::f64::consts::PI * size / (4 as f64)
+    std::f64::consts::PI * size / 4.0
 }
 
 fn assert_close(a:f64, b:f64, epsilon:f64) {
